Configure react router to use browser history
Uses the native browser history api instead of hash urls, resulting in a
cleaner experience. For browsers that do not support the history api,
the page is fully refreshed, which is why the frontend serving code is
changed to serve the index page for all urls. Serving the index at all
urls requires moving the api and dev-login boot scripts in front of the
frontend boot script, so that express favors their route definitions.