Fix bug that caused any edits done on the participant page to fail
Problem was that the behavior of the massAssign backend changed a bit when it was migrated from LoopBack REST to Express. Previously, LoopBack automatically wrapped the ids parameter to an array if it was an integer and the current endpoint doesn't. Also, LoopBack wrapped the result of massAssign in a "result" object, which also no longer happens. Now the frontend has been updated to use the proper conventions of the backend.

The functional tests didn't catch this since 1) they were never tested with non-array ids parameter 2) they never tested the return value of massAssign. We no longer need to support integer arguments for ids, which fixes (1). I also added some tests for the shape of massAssign's return value, preventing (2) from happening again. In the long run we should improve end-to-end test coverage.