Fix bug in participation dates mapping
If a payment type was not mapped, it would cause undefined values to be
injected into a participants participation date mappings. Those undefineds
then would be passed into moment, resulting in the current timestamp being
injected into the database for that user as a participation date.

Now those undefineds will be caught and a warning issued.