Simplify conf.js exports
The values were exported in a default-exported object, which is not a
very good practice (it emulates the old cjs exports in a way, but is
inefficient to use with es modules). The properties were also functions
that returned the actual values, but there wasn't any real benefit to
doing so, so the wrapping functions are now removed.